refactor MIDI Host callback and RX buffers
Use double-buffer so USB can run more freely.
Then callback gets a span of the last written buffer.
Which it can push to a FIFO or handle immediately (TBD).

Also removed need to MidiHost to be constructed with a host handle
which opens the way for us to create MidiHost in main()