Load module states before setting knobs
Fixes https://github.com/4ms/metamodule-plugin-examples/issues/6

MatrixMixer sets the paramQuantity->min to -1 in the constructor, then sets it to -1 or 0 based on the data in the json. When the knob's value is set by the patch player, the actual value stored depends on the paramQuantity->min. Therefore the initial value of the knob needs to be set after we load the data from json.